Change update_wishlist logic: leave item in a wishlist when logging in
The "like" button next to a product name both adds and removes the item
to/from the user's wishlist. When the user is logged out, all "like"
buttons are empty, i.e. the item is not in a wishlist because the
user is not logged in.

Previously clicking on this button when the user was logged out
triggered a logging in page after which the item was either added
or removed to/from the user's wishlist.

Now in this scenario the "like" button can only add an item to the
wishlist to avoid confusion.

Explanation: if I as a user, not being logged in but not aware of it,
look through the catalog, find an item that I like, click on an "empty
like" button, I would expect the item to be added to my wishlist,
not removed.